Cria handler para tratar exceções de serialização e desserialização
